work on fast regular expressions

+++ bin/mktest.sh
@@ -17,5 +17,14 @@ TOCUT=`expr 1 + ${#DEPDIR}`
 
 ROOT=`echo $FILEDIR | cut -c$TOCUT- | sed -e 's/\/[^/][^/]*/..\\\\\//g'`
 
-cat $DIR/test.html.template | sed -e "s/{CLASS}/$1/g" | sed -e "s/{ROOT}/$ROOT/g" > $FILEDIR/${FNAME}_test.html
-cat $DIR/test.js.template | sed -e "s/{CLASS}/$1/g" | sed -e "s/{ROOT}/$ROOT/g" > $FILEDIR/${FNAME}_test.js
+
+if [ -e $FILEDIR/${FNAME}_test.html ]; then
+    echo "test html already exists"
+else
+    cat $DIR/test.html.template | sed -e "s/{CLASS}/$1/g" | sed -e "s/{ROOT}/$ROOT/g" > $FILEDIR/${FNAME}_test.html
+fi
+if [ -e  $FILEDIR/${FNAME}_test.js ]; then
+    echo "test .js already exists"
+else
+    cat $DIR/test.js.template | sed -e "s/{CLASS}/$1/g" | sed -e "s/{ROOT}/$ROOT/g" > $FILEDIR/${FNAME}_test.js
+fi
